Implement refcounting of posix shared memory objects
Deletion of shmfs files occurs when both of the following
conditions are met: shm_unlink called on file, and all existing
mappings are unmapped.

Memory manager now notifies shm fs on first mmap and last munmap.
In the intervening time if shm_unlink is called on file, deletion
of file object is delayed until the last munmap.

Signed-off-by: Vikas Tikoo <vitikoo@microsoft.com>